third_party/blink/web_tests/external/wpt/html/infrastructure/common-dom-interfaces/collections/htmloptionscollection-expected.txt

This is a testharness.js-based test.
[FAIL] Insert <p><option>6</option></p> into <select>
  assert_equals: Correct length expected 5 but got 6
[FAIL] Insert <optgroup><optgroup><option>6</option></optgroup></optgroup> into <select>
  assert_equals: Correct length expected 5 but got 6
Harness: the test ran to completion.
